Vehicle mileage data service changes

+++ dotnet/corecomponents/net.atos.daf.ct2.vehicle/entity/dtoVehicleMileage.cs
@@ -8,9 +8,9 @@ namespace net.atos.daf.ct2.vehicle.entity
     {
         public int id { get; set; }
         public long evt_timestamp { get; set; }
-        public int odo_mileage { get; set; }
-        public int odo_distance { get; set; }
-        public int real_distance { get; set; }
+        public decimal odo_mileage { get; set; }
+        public decimal odo_distance { get; set; }
+        public decimal real_distance { get; set; }
         public string vin { get; set; }
         public long modified_at { get; set; }
     }
